Controller improvements, New Models, Better APIS, Better Math

+++ CSGAPI/Models/UserAccounts.cs
@@ -1,3 +1,4 @@
+using System;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
 
@@ -8,7 +9,7 @@ namespace CSGAPI.Models
         public string Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public double Balance { get; set; }
-        public double WithdrawalLimit { get; set; }
+        public Decimal Balance { get; set; }
+        public Decimal WithdrawalLimit { get; set; }
     }
 }
